Store build log in the database.
The idea is to make the logs available to non build machines.

This permit to setup a dedicated host to follow all the builder machines
with the web interface.

This is not done in realtime to avoid possible network/database flood.
The log display on the web interface of builders is still realtime.

The log is stored in a BLOBCol column at the end of the build process,
during send_build_log, which restrict usage to MySQL, PostgreSQL and
SQLite backends.

* rebuildd/RebuilddLog.py (Log): New object to store the log in the
  database.

* rebuildd.py (create_db): Create the "log" table, after the "job" one to
  get "ForeignKey" reference.

* rebuildd/Rebuildd.py: Import the new "Log" class.
  (Rebuildd.add_job): Create the "log" row with the
  reference to the current "job".

* rebuildd/Job.py: Import the new "Log" class.
  (Job): New "log" attribute.
  (Job.send_build_log): We need to dump the log file even
  if "mail_successful" and "mail_failed" are "False", move the dump at
  the beginning of the method.
  Store the log in the database.

* rebuildd/RebuilddHTTPServer.py (RequestJob.GET): Get the log from the
  database if the logfile is unreadable.

Closes: #657635