Log the error for unauthorized dist/arch (#570425)
When using "rebuildd-job", it will completely silently fail for
many different kinds of syntax or data errors.  For example, if an
architecture is specified which is not enabled in the config
file, it will accept the input and exit without doing anything.

The "rebuildd" tool itself also has several silent failures, one
of which will cause a Job object to get stuck in the "BUILDING"
state. If an exception occurs, such as "No such file or directory", the
Job will silently fail and get stuck in the "BUILDING" state.
(Kyle Moffett <Kyle.D.Moffett@boeing.com>)

Signed-off-by: Julien Danjou <julien@danjou.info>